Implementation of obtaining a sequence of frames using eeprom

+++ External/led_panels_driver/Inc/led_panels_driver.h
@@ -5,7 +5,7 @@
 #include "led_panels_defs.h"
 
 __attribute__((always_inline))
-inline static uint16_t get_side_size(uint16_t size)
+inline static uint8_t get_side_size(uint16_t size)
 {
   static const uint8_t sides_sizes[] = { 8U, 16U };
   return sides_sizes[size >> 8]; // size / 256
